Updated the Map
The Map now defaults to tracking the User's direction via gyroscope and magnetometer.
This behavior can be toggled off via the Compass indicator in the top right corner, however it's now always on by default.

The Map now allows to click on any of the default Markers, allowing to better see the name and the type of Marker of the place it belongs to.