Added Favor's UI in Map
Added FavorMarker, the UI element that displays a Favor on the Map.

Updated the Map, so that it refreshes whenever a new Favor is added.

Updated the New Favor Sheet (and New Favor Icon Sheet): it's now possible to customize some of the many parameters of a Favor, and to then add it to the list of available ones, so that it can be displayed on the Map