Added the Location Selector
Added the ability to choose a Location when creating a Favor!

The 'New Favor' sheet now has a preview of the selected Location in a small Map.

The button to select a location now opens a new sheet, consisting of another Map where the User can move the Favor's Marker, effectively changing its location.