Created LoginView
The login view welcome the user with the concept of the map and ask to provide the name, surname, which color want for profile and neighbourhood. The view is made to not show up after the first use of the app, but next the user can always change is name, ecc. from the profile (ModifyProfileView)