Added Reviews to StarsView
StarsView now displays Reviews as well.

If there are no reviews, nothing changes.
If there is 1, it will be displayed without the individual rating of that user, because it will be the same as the average rating, shown in big.

If there is 2 or more, it will display the individual rating as well.